// front-end/src/pages/creator/components/menu.interface.ts
/*
 * @Author: sunji 2025506282@qq.com
 * @Date: 2022-11-03 10:37:58
 * @LastEditors: sunji 2025506282@qq.com
 * @LastEditTime: 2022-11-03 10:40:56
 * @FilePath: \front-end\src\pages\creator\components\menu.interface.ts
 * @Description: 这是默认设置,请设置`customMade`, 打开koroFileHeader查看配置 进行设置: https://github.com/OBKoro1/koro1FileHeader/wiki/%E9%85%8D%E7%BD%AE
 */
export interface IMenuItem {
  label: string
  value: string
  icon?: string
  children?: IMenuItem[]
}